Moved Managers, and supporting classes, to a Manager folder/namespace
Implemented LogMonitor for Managers to be able to see what's going on with the server

LogMonitoringTraceListener was added to power LogMonitor

BdsManager now has several tracking/status properties so it knows things like if the server is running or not, last time it was started/stopped, last log on/off, etc.

BackupManager was implemented for taking online and offline backups manually.  Backup automation needs to be added.